+++ dflash/scripts/quantize_draft_q8.py
@@ -7,7 +7,7 @@ to Q8_0 (~50% size reduction vs BF16).  Norm weights stay F32
 (precision-critical, tiny).
 
 The output GGUF uses the same arch and tensor naming as
-convert_dflash_to_gguf.py so gguf_draft_loader.cpp can load it.
+convert_dflash_to_gguf.py so draft_gguf_loader.cpp can load it.
 
 Usage:
     python3 scripts/quantize_draft_q8.py \
