+++ dflash/test/smoke_draft_graph.cpp
@@ -14,7 +14,7 @@
 
 #include "dflash27b.h"
 #include "internal.h"
-#include "dflash_graph.h"
+#include "draft_graph.h"
 
 #include "ggml.h"
 #include "ggml-alloc.h"
