+++ dflash/test/test_vs_oracle.cpp
@@ -9,7 +9,7 @@
 
 #include "dflash27b.h"
 #include "internal.h"
-#include "dflash_graph.h"
+#include "draft_graph.h"
 
 #include "ggml.h"
 #include "ggml-alloc.h"
